requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.3
robotframework==7.3
robotframework-requests==0.9.7
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.4.0
